restructure + fix button alignment
