+++ frontend/src/products/utils.ts
@@ -1,0 +1,2 @@
+export const convertToDisplayPrice = (price: number) =>
+  (Math.round(price * 100) / 100).toFixed(2);
